Added Unit Tests for Mileage Form
Mileage form is now complete and passes all tests.
